Protect RestApi via masterkey Header-Attribute

+++ src/main/java/at/fhstp/bis19/prog4/snowdogs/sniffable/controller/GenericContoller.java
@@ -3,44 +3,70 @@ package at.fhstp.bis19.prog4.snowdogs.sniffable.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.env.Environment;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 import at.fhstp.bis19.prog4.snowdogs.sniffable.entity.BaseEntity;
 import at.fhstp.bis19.prog4.snowdogs.sniffable.repo.BaseCrudRepository;
 
 public class GenericContoller<T extends BaseEntity> {
+	private final String masterkeyHeaderAttribute = "masterkey";
+	private final String masterkeyConfigProperty = "sniffers.restapi.masterKey";
+	
+	@Autowired
+	private Environment env;
+
 	@Autowired
 	private BaseCrudRepository<T> repo;
-	
-	
+
 	@GetMapping
-	public List<T> read() {
-		return (List<T>) repo.findAll();
+	public List<T> read(@RequestHeader(masterkeyHeaderAttribute) String headerKey) {
+		String mk = env.getProperty(masterkeyConfigProperty);
+		if (mk != null && mk.equals(headerKey)) {
+			return (List<T>) repo.findAll();
+		}
+		return null;
 	}
-	
+
 	@GetMapping(value = "{id}")
-	public T readByID(@PathVariable(value = "id") int id) {
-		return repo.findById(id).orElse(null);
+	public T readByID(@PathVariable(value = "id") int id, @RequestHeader(masterkeyHeaderAttribute) String headerKey) {
+		String mk = env.getProperty(masterkeyConfigProperty);
+		if (mk != null && mk.equals(headerKey)) {
+			return repo.findById(id).orElse(null);
+		}
+		return null;
 	}
-	
+
 	@PostMapping
-	public T create(@RequestBody T entity) {
-		return repo.save(entity);
+	public T create(@RequestBody T entity, @RequestHeader("masterkey") String headerKey) {
+		String mk = env.getProperty(masterkeyConfigProperty);
+		if (mk != null && mk.equals(headerKey)) {
+			return repo.save(entity);
+		}
+		return null;
 	}
-	
+
 	@PutMapping(value = "{id}")
-	public T update(@PathVariable(value = "id") int id, @RequestBody T entity) {
-		return repo.save(entity);
+	public T update(@PathVariable(value = "id") int id, @RequestBody T entity, @RequestHeader(masterkeyHeaderAttribute) String headerKey) {
+		String mk = env.getProperty(masterkeyConfigProperty);
+		if (mk != null && mk.equals(headerKey)) {
+			return repo.save(entity);
+		}
+		return null;
 	}
-	
+
 	@DeleteMapping(value = "{id}")
-	public void delete(@PathVariable(value = "id") int id) {
-		repo.deleteById(id);
+	public void delete(@PathVariable(value = "id") int id, @RequestHeader(masterkeyHeaderAttribute) String headerKey) {
+		String mk = env.getProperty(masterkeyConfigProperty);
+		if (mk != null && mk.equals(headerKey)) {
+			repo.deleteById(id);
+		}
 	}
-	
+
 }
